plotting w/ dates (trend)
