+++ libs/utils/__init__.py
@@ -5,7 +5,7 @@ from .data import download_data, download_data_indexes
 from .error_handler import has_critical_error
 
 from .formatting import name_parser, index_extractor, fund_list_extractor, index_appender
-from .formatting import dates_extractor_list, date_extractor, get_daterange
+from .formatting import dates_extractor_list, date_extractor, get_daterange, dates_convert_from_index
 
 from .plotting import dual_plotting, generic_plotting, histogram, bar_chart, specialty_plotting, shape_plotting
 from .plotting import candlestick
