+++ libs/utils/formatting.py
@@ -124,3 +124,14 @@ def period_parser(period: str):
     else:
         return 1, 0, 0
 
+
+def dates_convert_from_index(df: pd.DataFrame, list_of_xlists: list) -> list:
+    new_l_of_xls = []
+    if len(list_of_xlists) > 0:
+        for xlist in list_of_xlists:
+            new_xlist = []
+            for x in xlist:
+                new_xlist.append(df.index[x])
+            new_l_of_xls.append(new_xlist)
+    return new_l_of_xls
+
